+++ core/station/impl/src/models/request.rs
@@ -1,7 +1,7 @@
-use super::request_policy_rule::{RequestEvaluationResult, RequestPolicyRuleInput};
+use super::request_policy_rule::RequestEvaluationResult;
 use super::{
-    ChangeAssets, ConfigureExternalCanisterOperationKind, DisplayUser, EvaluationStatus,
-    RequestApproval, RequestApprovalStatus, RequestOperation, RequestStatus, UserId, UserKey,
+    DisplayUser, EvaluationStatus, RequestApproval, RequestApprovalStatus, RequestOperation,
+    RequestStatus, UserId, UserKey,
 };
 use crate::core::evaluation::{
     Evaluate, REQUEST_APPROVE_RIGHTS_REQUEST_POLICY_RULE_EVALUATOR, REQUEST_POLICY_RULE_EVALUATOR,
@@ -12,15 +12,10 @@ use crate::core::ic_cdk::next_time;
 use crate::core::request::{
     RequestApprovalRightsEvaluator, RequestEvaluator, RequestPossibleApproversFinder,
 };
-use crate::core::validation::{
-    EnsureAccount, EnsureAddressBookEntry, EnsureAsset, EnsureIdExists, EnsureNamedRule,
-    EnsureRequestPolicy, EnsureUser, EnsureUserGroup,
-};
-use crate::errors::{EvaluateError, RequestError, ValidationError};
-use crate::models::resource::{ExecutionMethodResourceTarget, ValidationMethodResourceTarget};
+use crate::errors::{EvaluateError, RequestError};
 use crate::repositories::USER_REPOSITORY;
 use candid::{CandidType, Deserialize};
-use orbit_essentials::model::{ContextualModel, ModelKey};
+use orbit_essentials::model::ModelKey;
 use orbit_essentials::repository::Repository;
 use orbit_essentials::storable;
 use orbit_essentials::{
@@ -191,169 +186,6 @@ fn validate_requested_by(requested_by: &UserId) -> ModelValidatorResult<RequestE
     Ok(())
 }
 
-fn validate_request_operation_foreign_keys(
-    operation: &RequestOperation,
-) -> ModelValidatorResult<ValidationError> {
-    match operation {
-        RequestOperation::ManageSystemInfo(_) => (),
-        RequestOperation::Transfer(op) => {
-            EnsureAccount::id_exists(&op.input.from_account_id)?;
-            EnsureAsset::id_exists(&op.input.from_asset_id)?;
-        }
-        RequestOperation::AddAccount(op) => {
-            op.input.read_permission.validate()?;
-            op.input.configs_permission.validate()?;
-            op.input.transfer_permission.validate()?;
-
-            if let Some(policy_rule) = &op.input.transfer_request_policy {
-                policy_rule.validate()?;
-            }
-
-            if let Some(policy_rule) = &op.input.configs_request_policy {
-                policy_rule.validate()?;
-            }
-        }
-        RequestOperation::EditAccount(op) => {
-            EnsureAccount::id_exists(&op.input.account_id)?;
-
-            if let Some(allow) = &op.input.read_permission {
-                allow.validate()?;
-            }
-
-            if let Some(allow) = &op.input.configs_permission {
-                allow.validate()?;
-            }
-
-            if let Some(allow) = &op.input.transfer_permission {
-                allow.validate()?;
-            }
-
-            if let Some(RequestPolicyRuleInput::Set(criteria)) = &op.input.configs_request_policy {
-                criteria.validate()?;
-            }
-
-            if let Some(RequestPolicyRuleInput::Set(policy_rule)) =
-                &op.input.transfer_request_policy
-            {
-                policy_rule.validate()?;
-            }
-
-            if let Some(ChangeAssets::ReplaceWith { assets }) = &op.input.change_assets {
-                EnsureAsset::id_list_exists(assets)?;
-            }
-
-            if let Some(ChangeAssets::Change {
-                add_assets,
-                remove_assets,
-            }) = &op.input.change_assets
-            {
-                EnsureAsset::id_list_exists(add_assets)?;
-                EnsureAsset::id_list_exists(remove_assets)?;
-            }
-        }
-        RequestOperation::AddAddressBookEntry(_) => (),
-        RequestOperation::EditAddressBookEntry(op) => {
-            EnsureAddressBookEntry::id_exists(&op.input.address_book_entry_id)?;
-        }
-        RequestOperation::RemoveAddressBookEntry(op) => {
-            EnsureAddressBookEntry::id_exists(&op.input.address_book_entry_id)?;
-        }
-        RequestOperation::AddUser(op) => {
-            EnsureUserGroup::id_list_exists(&op.input.groups)?;
-        }
-        RequestOperation::EditUser(op) => {
-            EnsureUser::id_exists(&op.input.user_id)?;
-
-            if let Some(group_ids) = &op.input.groups {
-                EnsureUserGroup::id_list_exists(group_ids)?;
-            }
-        }
-        RequestOperation::EditPermission(op) => {
-            op.input.resource.validate()?;
-
-            if let Some(user_ids) = &op.input.users {
-                EnsureUser::id_list_exists(user_ids)?;
-            }
-
-            if let Some(group_ids) = &op.input.user_groups {
-                EnsureUserGroup::id_list_exists(group_ids)?;
-            }
-        }
-        RequestOperation::AddUserGroup(_) => (),
-        RequestOperation::EditUserGroup(op) => {
-            EnsureUserGroup::id_exists(&op.input.user_group_id)?;
-        }
-        RequestOperation::RemoveUserGroup(ok) => {
-            EnsureUserGroup::id_exists(&ok.input.user_group_id)?;
-        }
-        RequestOperation::SystemUpgrade(_) => (),
-        RequestOperation::ChangeExternalCanister(_) => (),
-        RequestOperation::ConfigureExternalCanister(op) => {
-            let canister_id = op.canister_id;
-            if let ConfigureExternalCanisterOperationKind::Settings(settings) = &op.kind {
-                if let Some(updated_request_policies) = &settings.request_policies {
-                    ContextualModel::new(updated_request_policies.clone(), canister_id)
-                        .validate()?;
-                }
-            }
-        }
-        RequestOperation::FundExternalCanister(_) => (),
-        RequestOperation::MonitorExternalCanister(_) => (),
-        RequestOperation::CreateExternalCanister(op) => {
-            op.input.validate()?;
-        }
-        RequestOperation::CallExternalCanister(op) => {
-            let validation_method_target: ValidationMethodResourceTarget =
-                op.input.validation_method.clone().into();
-            validation_method_target.validate()?;
-            let execution_method_target: ExecutionMethodResourceTarget =
-                op.input.execution_method.clone().into();
-            execution_method_target.validate()?;
-        }
-        RequestOperation::SnapshotExternalCanister(_) => (),
-        RequestOperation::RestoreExternalCanister(_) => (),
-        RequestOperation::PruneExternalCanister(_) => (),
-        RequestOperation::AddRequestPolicy(op) => {
-            op.input.specifier.validate()?;
-            op.input.rule.validate()?;
-        }
-        RequestOperation::EditRequestPolicy(op) => {
-            EnsureRequestPolicy::id_exists(&op.input.policy_id)?;
-
-            if let Some(specifier) = &op.input.specifier {
-                specifier.validate()?;
-            }
-
-            if let Some(policy_rule) = &op.input.rule {
-                policy_rule.validate()?;
-            }
-        }
-        RequestOperation::RemoveRequestPolicy(op) => {
-            EnsureRequestPolicy::id_exists(&op.input.policy_id)?;
-        }
-        RequestOperation::SetDisasterRecovery(op) => {
-            if let Some(committee) = &op.input.committee {
-                EnsureUserGroup::id_exists(&committee.user_group_id)?;
-            }
-        }
-        RequestOperation::AddAsset(_) => (),
-        RequestOperation::EditAsset(op) => {
-            EnsureAsset::id_exists(&op.input.asset_id)?;
-        }
-        RequestOperation::RemoveAsset(op) => {
-            EnsureAsset::id_exists(&op.input.asset_id)?;
-        }
-        RequestOperation::AddNamedRule(_) => (),
-        RequestOperation::EditNamedRule(op) => {
-            EnsureNamedRule::id_exists(&op.input.named_rule_id)?;
-        }
-        RequestOperation::RemoveNamedRule(op) => {
-            EnsureNamedRule::id_exists(&op.input.named_rule_id)?;
-        }
-    }
-    Ok(())
-}
-
 impl ModelValidator<RequestError> for Request {
     fn validate(&self) -> ModelValidatorResult<RequestError> {
         validate_title(&self.title)?;
@@ -362,7 +194,7 @@ impl ModelValidator<RequestError> for Request {
         validate_expiration_dt(&self.expiration_dt)?;
         validate_execution_plan(&self.execution_plan)?;
         validate_status(&self.status)?;
-        validate_request_operation_foreign_keys(&self.operation)?;
+        self.operation.validate()?;
 
         Ok(())
     }
@@ -515,17 +347,6 @@ impl Request {
 
 #[cfg(test)]
 mod tests {
-    use crate::core::validation::disable_mock_resource_validation;
-    use crate::models::asset_test_utils::mock_asset;
-    use crate::models::permission::Allow;
-    use crate::models::{
-        Account, AccountKey, AddAccountOperationInput, AddAssetOperationInput, AddUserOperation,
-        AddUserOperationInput, Blockchain, Metadata, TokenStandard, TransferOperation,
-        TransferOperationInput,
-    };
-    use crate::repositories::ACCOUNT_REPOSITORY;
-    use crate::services::{AccountService, AssetService};
-
     use super::request_test_utils::mock_request;
     use super::*;
 
@@ -687,269 +508,6 @@ mod tests {
         let result = validate_summary(&request.summary);
 
         assert!(result.is_ok());
-    }
-
-    #[tokio::test]
-    async fn test_request_operation_is_valid() {
-        disable_mock_resource_validation();
-
-        let asset = AssetService::default()
-            .create(
-                AddAssetOperationInput {
-                    name: "a".to_owned(),
-                    symbol: "a".to_owned(),
-                    decimals: 0,
-                    metadata: Metadata::default(),
-                    blockchain: Blockchain::InternetComputer,
-                    standards: vec![TokenStandard::InternetComputerNative],
-                },
-                None,
-            )
-            .expect("Failed to create asset");
-
-        let account_service = AccountService::default();
-        let account = account_service
-            .create_account(
-                AddAccountOperationInput {
-                    name: "a".to_owned(),
-                    assets: vec![asset.id],
-                    metadata: Metadata::default(),
-                    read_permission: Allow::default(),
-                    configs_permission: Allow::default(),
-                    transfer_permission: Allow::default(),
-                    configs_request_policy: None,
-                    transfer_request_policy: None,
-                },
-                None,
-            )
-            .await
-            .expect("Failed to create account");
-
-        let operation = RequestOperation::Transfer(TransferOperation {
-            transfer_id: None,
-            fee: None,
-
-            input: TransferOperationInput {
-                network: "mainnet".to_string(),
-                amount: 1u64.into(),
-                fee: None,
-                metadata: Metadata::default(),
-                to: "0x1234".to_string(),
-                from_account_id: account.id,
-                from_asset_id: asset.id,
-                with_standard: TokenStandard::InternetComputerNative,
-            },
-            asset,
-        });
-
-        let result = validate_request_operation_foreign_keys(&operation);
-
-        assert!(result.is_ok());
-    }
-
-    #[tokio::test]
-    async fn fail_request_operation_with_invalid_id() {
-        disable_mock_resource_validation();
-
-        validate_request_operation_foreign_keys(&RequestOperation::Transfer(TransferOperation {
-            transfer_id: None,
-            fee: None,
-            input: TransferOperationInput {
-                network: "mainnet".to_string(),
-                amount: 1u64.into(),
-                fee: None,
-                metadata: Metadata::default(),
-                to: "0x1234".to_string(),
-                from_account_id: [0; 16],
-                from_asset_id: [0; 16],
-                with_standard: TokenStandard::InternetComputerNative,
-            },
-            asset: mock_asset(),
-        }))
-        .expect_err("Invalid account id should fail");
-
-        validate_request_operation_foreign_keys(&RequestOperation::AddUser(AddUserOperation {
-            user_id: None,
-            input: AddUserOperationInput {
-                name: "user-1".to_string(),
-                identities: vec![],
-                groups: vec![[1; 16]],
-                status: crate::models::UserStatus::Active,
-            },
-        }))
-        .expect_err("Invalid user group id should fail");
-
-        validate_request_operation_foreign_keys(&RequestOperation::EditUserGroup(
-            crate::models::EditUserGroupOperation {
-                input: crate::models::EditUserGroupOperationInput {
-                    user_group_id: [0; 16],
-                    name: "a".to_owned(),
-                },
-            },
-        ))
-        .expect_err("Invalid user group id should fail");
-        validate_request_operation_foreign_keys(&RequestOperation::RemoveUserGroup(
-            crate::models::RemoveUserGroupOperation {
-                input: crate::models::RemoveUserGroupOperationInput {
-                    user_group_id: [0; 16],
-                },
-            },
-        ))
-        .expect_err("Invalid user group id should fail");
-
-        validate_request_operation_foreign_keys(&RequestOperation::AddRequestPolicy(
-            crate::models::AddRequestPolicyOperation {
-                policy_id: None,
-                input: crate::models::AddRequestPolicyOperationInput {
-                    specifier: crate::models::request_specifier::RequestSpecifier::EditUser(
-                        crate::models::resource::ResourceIds::Ids(vec![[1; 16]]),
-                    ),
-                    rule: crate::models::request_policy_rule::RequestPolicyRule::AutoApproved,
-                },
-            },
-        ))
-        .expect_err("Invalid request specifier should fail");
-
-        validate_request_operation_foreign_keys(&RequestOperation::EditRequestPolicy(
-            crate::models::EditRequestPolicyOperation {
-                input: crate::models::EditRequestPolicyOperationInput {
-                    policy_id: [0; 16],
-                    specifier: None,
-                    rule: None,
-                },
-            },
-        ))
-        .expect_err("Invalid request policy id should fail");
-
-        validate_request_operation_foreign_keys(&RequestOperation::RemoveRequestPolicy(
-            crate::models::RemoveRequestPolicyOperation {
-                input: crate::models::RemoveRequestPolicyOperationInput { policy_id: [0; 16] },
-            },
-        ))
-        .expect_err("Invalid request policy id should fail");
-
-        validate_request_operation_foreign_keys(&RequestOperation::AddAccount(
-            crate::models::AddAccountOperation {
-                account_id: None,
-                input: crate::models::AddAccountOperationInput {
-                    name: "a".to_owned(),
-                    assets: vec![],
-                    metadata: Metadata::default(),
-                    read_permission: Allow {
-                        auth_scope: crate::models::permission::AuthScope::Restricted,
-                        users: vec![[1; 16]],
-                        user_groups: vec![],
-                    },
-                    configs_permission: Allow::default(),
-                    transfer_permission: Allow::default(),
-                    configs_request_policy: None,
-                    transfer_request_policy: None,
-                },
-            },
-        ))
-        .expect_err("Invalid user id should fail");
-
-        validate_request_operation_foreign_keys(&RequestOperation::EditAccount(
-            crate::models::EditAccountOperation {
-                input: crate::models::EditAccountOperationInput {
-                    account_id: [0; 16],
-                    change_assets: None,
-                    read_permission: None,
-                    configs_permission: None,
-                    transfer_permission: None,
-                    configs_request_policy: None,
-                    transfer_request_policy: None,
-                    name: None,
-                },
-            },
-        ))
-        .expect_err("Invalid account id should fail");
-
-        ACCOUNT_REPOSITORY.insert(
-            AccountKey { id: [0; 16] },
-            Account {
-                id: [0; 16],
-                name: "a".to_owned(),
-                seed: [0; 16],
-                assets: vec![],
-                addresses: vec![],
-                metadata: Metadata::default(),
-                transfer_request_policy_id: None,
-                configs_request_policy_id: None,
-                last_modification_timestamp: 0,
-            },
-        );
-
-        validate_request_operation_foreign_keys(&RequestOperation::EditAccount(
-            crate::models::EditAccountOperation {
-                input: crate::models::EditAccountOperationInput {
-                    account_id: [0; 16],
-                    change_assets: Some(ChangeAssets::ReplaceWith {
-                        assets: vec![[0; 16]],
-                    }),
-                    read_permission: None,
-                    configs_permission: None,
-                    transfer_permission: None,
-                    configs_request_policy: None,
-                    transfer_request_policy: None,
-                    name: None,
-                },
-            },
-        ))
-        .expect_err("Invalid asset id should fail");
-
-        ACCOUNT_REPOSITORY.clear();
-
-        validate_request_operation_foreign_keys(&RequestOperation::EditAddressBookEntry(
-            crate::models::EditAddressBookEntryOperation {
-                input: crate::models::EditAddressBookEntryOperationInput {
-                    address_book_entry_id: [0; 16],
-                    address_owner: None,
-                    change_metadata: None,
-                    labels: None,
-                },
-            },
-        ))
-        .expect_err("Invalid address book entry id should fail");
-
-        validate_request_operation_foreign_keys(&RequestOperation::RemoveAddressBookEntry(
-            crate::models::RemoveAddressBookEntryOperation {
-                input: crate::models::RemoveAddressBookEntryOperationInput {
-                    address_book_entry_id: [0; 16],
-                },
-            },
-        ))
-        .expect_err("Invalid address book entry id should fail");
-
-        validate_request_operation_foreign_keys(&RequestOperation::EditUser(
-            crate::models::EditUserOperation {
-                input: crate::models::EditUserOperationInput {
-                    user_id: [0; 16],
-                    groups: None,
-                    name: None,
-                    identities: None,
-                    status: None,
-                    cancel_pending_requests: None,
-                },
-            },
-        ))
-        .expect_err("Invalid user id should fail");
-
-        validate_request_operation_foreign_keys(&RequestOperation::EditPermission(
-            crate::models::EditPermissionOperation {
-                input: crate::models::EditPermissionOperationInput {
-                    resource: crate::models::resource::Resource::Account(
-                        crate::models::resource::AccountResourceAction::Read(
-                            crate::models::resource::ResourceId::Id([0; 16]),
-                        ),
-                    ),
-                    users: None,
-                    user_groups: None,
-                    auth_scope: None,
-                },
-            },
-        ))
-        .expect_err("Invalid resource id should fail");
     }
 }
