+++ core/station/impl/src/models/request_operation.rs
@@ -8,13 +8,17 @@ use super::{
     ExternalCanisterState, MetadataItem, NamedRuleId, TokenStandard, UserGroupId, UserId,
     UserStatus,
 };
-use crate::core::validation::EnsureExternalCanister;
-use crate::errors::ValidationError;
+use crate::core::validation::{
+    EnsureAccount, EnsureAddressBookEntry, EnsureAsset, EnsureExternalCanister, EnsureIdExists,
+    EnsureNamedRule, EnsureRequestPolicy, EnsureUser, EnsureUserGroup,
+};
+use crate::errors::{ExternalCanisterValidationError, ValidationError};
+use crate::models::resource::ExecutionMethodResourceTarget;
 use crate::models::Metadata;
 use candid::Principal;
 use orbit_essentials::cdk::api::management_canister::main::{self as mgmt};
 use orbit_essentials::cmc::SubnetSelection;
-use orbit_essentials::model::{ModelValidator, ModelValidatorResult};
+use orbit_essentials::model::{ContextualModel, ModelValidator, ModelValidatorResult};
 use orbit_essentials::{storable, types::UUID};
 use std::{collections::HashSet, fmt::Display};
 
@@ -783,6 +787,29 @@ pub struct CallExternalCanisterOperationInput {
     pub execution_method_cycles: Option<u64>,
 }
 
+impl ModelValidator<ValidationError> for CallExternalCanisterOperationInput {
+    fn validate(&self) -> ModelValidatorResult<ValidationError> {
+        // Disallows the wildcard execution method
+        if self.execution_method.method_name == CanisterMethod::WILDCARD {
+            return Err(ValidationError::ExternalCanisterValidationError(
+                ExternalCanisterValidationError::ValidationError {
+                    info: "Wildcard execution method is not allowed.".to_string(),
+                },
+            ));
+        }
+
+        // Validate both methods
+        let validation_method_target: ValidationMethodResourceTarget =
+            self.validation_method.clone().into();
+        validation_method_target.validate()?;
+        let execution_method_target: ExecutionMethodResourceTarget =
+            self.execution_method.clone().into();
+        execution_method_target.validate()?;
+
+        Ok(())
+    }
+}
+
 #[storable]
 #[derive(Clone, Debug, PartialEq, Eq, Hash, PartialOrd, Ord)]
 pub struct CallExternalCanisterOperation {
@@ -969,9 +996,201 @@ pub struct EditNamedRuleOperationInput {
     pub rule: Option<RequestPolicyRule>,
 }
 
+impl ModelValidator<ValidationError> for RequestOperation {
+    fn validate(&self) -> ModelValidatorResult<ValidationError> {
+        match self {
+            RequestOperation::ManageSystemInfo(_) => (),
+            RequestOperation::Transfer(op) => {
+                EnsureAccount::id_exists(&op.input.from_account_id)?;
+                EnsureAsset::id_exists(&op.input.from_asset_id)?;
+            }
+            RequestOperation::AddAccount(op) => {
+                op.input.read_permission.validate()?;
+                op.input.configs_permission.validate()?;
+                op.input.transfer_permission.validate()?;
+
+                if let Some(policy_rule) = &op.input.transfer_request_policy {
+                    policy_rule.validate()?;
+                }
+
+                if let Some(policy_rule) = &op.input.configs_request_policy {
+                    policy_rule.validate()?;
+                }
+            }
+            RequestOperation::EditAccount(op) => {
+                EnsureAccount::id_exists(&op.input.account_id)?;
+
+                if let Some(allow) = &op.input.read_permission {
+                    allow.validate()?;
+                }
+
+                if let Some(allow) = &op.input.configs_permission {
+                    allow.validate()?;
+                }
+
+                if let Some(allow) = &op.input.transfer_permission {
+                    allow.validate()?;
+                }
+
+                if let Some(RequestPolicyRuleInput::Set(criteria)) =
+                    &op.input.configs_request_policy
+                {
+                    criteria.validate()?;
+                }
+
+                if let Some(RequestPolicyRuleInput::Set(policy_rule)) =
+                    &op.input.transfer_request_policy
+                {
+                    policy_rule.validate()?;
+                }
+
+                if let Some(ChangeAssets::ReplaceWith { assets }) = &op.input.change_assets {
+                    EnsureAsset::id_list_exists(assets)?;
+                }
+
+                if let Some(ChangeAssets::Change {
+                    add_assets,
+                    remove_assets,
+                }) = &op.input.change_assets
+                {
+                    EnsureAsset::id_list_exists(add_assets)?;
+                    EnsureAsset::id_list_exists(remove_assets)?;
+                }
+            }
+            RequestOperation::AddAddressBookEntry(_) => (),
+            RequestOperation::EditAddressBookEntry(op) => {
+                EnsureAddressBookEntry::id_exists(&op.input.address_book_entry_id)?;
+            }
+            RequestOperation::RemoveAddressBookEntry(op) => {
+                EnsureAddressBookEntry::id_exists(&op.input.address_book_entry_id)?;
+            }
+            RequestOperation::AddUser(op) => {
+                EnsureUserGroup::id_list_exists(&op.input.groups)?;
+            }
+            RequestOperation::EditUser(op) => {
+                EnsureUser::id_exists(&op.input.user_id)?;
+
+                if let Some(group_ids) = &op.input.groups {
+                    EnsureUserGroup::id_list_exists(group_ids)?;
+                }
+            }
+            RequestOperation::EditPermission(op) => {
+                op.input.resource.validate()?;
+
+                if let Some(user_ids) = &op.input.users {
+                    EnsureUser::id_list_exists(user_ids)?;
+                }
+
+                if let Some(group_ids) = &op.input.user_groups {
+                    EnsureUserGroup::id_list_exists(group_ids)?;
+                }
+            }
+            RequestOperation::AddUserGroup(_) => (),
+            RequestOperation::EditUserGroup(op) => {
+                EnsureUserGroup::id_exists(&op.input.user_group_id)?;
+            }
+            RequestOperation::RemoveUserGroup(ok) => {
+                EnsureUserGroup::id_exists(&ok.input.user_group_id)?;
+            }
+            RequestOperation::SystemUpgrade(_) => (),
+            RequestOperation::ChangeExternalCanister(op) => {
+                let canister_id = op.input.canister_id;
+                EnsureExternalCanister::is_external_canister(canister_id)?;
+            }
+            RequestOperation::ConfigureExternalCanister(op) => {
+                let canister_id = op.canister_id;
+                EnsureExternalCanister::is_external_canister(canister_id)?;
+                if let ConfigureExternalCanisterOperationKind::Settings(settings) = &op.kind {
+                    if let Some(updated_request_policies) = &settings.request_policies {
+                        ContextualModel::new(updated_request_policies.clone(), canister_id)
+                            .validate()?;
+                    }
+                }
+            }
+            RequestOperation::FundExternalCanister(op) => {
+                let canister_id = op.canister_id;
+                EnsureExternalCanister::is_external_canister(canister_id)?;
+            }
+            RequestOperation::MonitorExternalCanister(op) => {
+                let canister_id = op.canister_id;
+                EnsureExternalCanister::is_external_canister(canister_id)?;
+            }
+            RequestOperation::CreateExternalCanister(op) => {
+                op.input.validate()?;
+            }
+            RequestOperation::CallExternalCanister(op) => {
+                op.input.validate()?;
+            }
+            RequestOperation::SnapshotExternalCanister(op) => {
+                let canister_id = op.input.canister_id;
+                EnsureExternalCanister::is_external_canister(canister_id)?;
+            }
+            RequestOperation::RestoreExternalCanister(op) => {
+                let canister_id = op.input.canister_id;
+                EnsureExternalCanister::is_external_canister(canister_id)?;
+            }
+            RequestOperation::PruneExternalCanister(op) => {
+                let canister_id = op.input.canister_id;
+                EnsureExternalCanister::is_external_canister(canister_id)?;
+            }
+            RequestOperation::AddRequestPolicy(op) => {
+                op.input.specifier.validate()?;
+                op.input.rule.validate()?;
+            }
+            RequestOperation::EditRequestPolicy(op) => {
+                EnsureRequestPolicy::id_exists(&op.input.policy_id)?;
+
+                if let Some(specifier) = &op.input.specifier {
+                    specifier.validate()?;
+                }
+
+                if let Some(policy_rule) = &op.input.rule {
+                    policy_rule.validate()?;
+                }
+            }
+            RequestOperation::RemoveRequestPolicy(op) => {
+                EnsureRequestPolicy::id_exists(&op.input.policy_id)?;
+            }
+            RequestOperation::SetDisasterRecovery(op) => {
+                if let Some(committee) = &op.input.committee {
+                    EnsureUserGroup::id_exists(&committee.user_group_id)?;
+                }
+            }
+            RequestOperation::AddAsset(_) => (),
+            RequestOperation::EditAsset(op) => {
+                EnsureAsset::id_exists(&op.input.asset_id)?;
+            }
+            RequestOperation::RemoveAsset(op) => {
+                EnsureAsset::id_exists(&op.input.asset_id)?;
+            }
+            RequestOperation::AddNamedRule(_) => (),
+            RequestOperation::EditNamedRule(op) => {
+                EnsureNamedRule::id_exists(&op.input.named_rule_id)?;
+            }
+            RequestOperation::RemoveNamedRule(op) => {
+                EnsureNamedRule::id_exists(&op.input.named_rule_id)?;
+            }
+        }
+        Ok(())
+    }
+}
+
 #[cfg(test)]
 mod test {
-    use crate::models::AccountAsset;
+    use super::*;
+    use crate::core::validation::disable_mock_resource_validation;
+    use crate::core::write_system_info;
+    use crate::errors::ExternalCanisterValidationError;
+    use crate::models::asset_test_utils::mock_asset;
+    use crate::models::permission::Allow;
+    use crate::models::{
+        Account, AccountKey, AddAccountOperationInput, AddAssetOperationInput, AddUserOperation,
+        AddUserOperationInput, Blockchain, Metadata, SystemInfo, TokenStandard, TransferOperation,
+        TransferOperationInput,
+    };
+    use crate::repositories::ACCOUNT_REPOSITORY;
+    use crate::services::{AccountService, AssetService};
+    use orbit_essentials::repository::Repository;
 
     use super::ChangeAssets;
 
@@ -1005,5 +1224,484 @@ mod test {
         assert!(assets.iter().any(|a| a.asset_id == [3; 16]));
 
         assert_eq!(assets.iter().filter(|a| a.asset_id == [3; 16]).count(), 1);
+    }
+
+    #[tokio::test]
+    async fn test_request_operation_is_valid() {
+        disable_mock_resource_validation();
+
+        let asset = AssetService::default()
+            .create(
+                AddAssetOperationInput {
+                    name: "a".to_owned(),
+                    symbol: "a".to_owned(),
+                    decimals: 0,
+                    metadata: Metadata::default(),
+                    blockchain: Blockchain::InternetComputer,
+                    standards: vec![TokenStandard::InternetComputerNative],
+                },
+                None,
+            )
+            .expect("Failed to create asset");
+
+        let account_service = AccountService::default();
+        let account = account_service
+            .create_account(
+                AddAccountOperationInput {
+                    name: "a".to_owned(),
+                    assets: vec![asset.id],
+                    metadata: Metadata::default(),
+                    read_permission: Allow::default(),
+                    configs_permission: Allow::default(),
+                    transfer_permission: Allow::default(),
+                    configs_request_policy: None,
+                    transfer_request_policy: None,
+                },
+                None,
+            )
+            .await
+            .expect("Failed to create account");
+
+        let operation = RequestOperation::Transfer(TransferOperation {
+            transfer_id: None,
+            fee: None,
+
+            input: TransferOperationInput {
+                network: "mainnet".to_string(),
+                amount: 1u64.into(),
+                fee: None,
+                metadata: Metadata::default(),
+                to: "0x1234".to_string(),
+                from_account_id: account.id,
+                from_asset_id: asset.id,
+                with_standard: TokenStandard::InternetComputerNative,
+            },
+            asset,
+        });
+
+        assert!(operation.validate().is_ok());
+    }
+
+    #[tokio::test]
+    async fn fail_request_operation_with_invalid_id() {
+        disable_mock_resource_validation();
+
+        RequestOperation::Transfer(TransferOperation {
+            transfer_id: None,
+            fee: None,
+            input: TransferOperationInput {
+                network: "mainnet".to_string(),
+                amount: 1u64.into(),
+                fee: None,
+                metadata: Metadata::default(),
+                to: "0x1234".to_string(),
+                from_account_id: [0; 16],
+                from_asset_id: [0; 16],
+                with_standard: TokenStandard::InternetComputerNative,
+            },
+            asset: mock_asset(),
+        })
+        .validate()
+        .expect_err("Invalid account id should fail");
+
+        RequestOperation::AddUser(AddUserOperation {
+            user_id: None,
+            input: AddUserOperationInput {
+                name: "user-1".to_string(),
+                identities: vec![],
+                groups: vec![[1; 16]],
+                status: crate::models::UserStatus::Active,
+            },
+        })
+        .validate()
+        .expect_err("Invalid user group id should fail");
+
+        RequestOperation::EditUserGroup(crate::models::EditUserGroupOperation {
+            input: crate::models::EditUserGroupOperationInput {
+                user_group_id: [0; 16],
+                name: "a".to_owned(),
+            },
+        })
+        .validate()
+        .expect_err("Invalid user group id should fail");
+        RequestOperation::RemoveUserGroup(crate::models::RemoveUserGroupOperation {
+            input: crate::models::RemoveUserGroupOperationInput {
+                user_group_id: [0; 16],
+            },
+        })
+        .validate()
+        .expect_err("Invalid user group id should fail");
+
+        RequestOperation::AddRequestPolicy(crate::models::AddRequestPolicyOperation {
+            policy_id: None,
+            input: crate::models::AddRequestPolicyOperationInput {
+                specifier: crate::models::request_specifier::RequestSpecifier::EditUser(
+                    crate::models::resource::ResourceIds::Ids(vec![[1; 16]]),
+                ),
+                rule: crate::models::request_policy_rule::RequestPolicyRule::AutoApproved,
+            },
+        })
+        .validate()
+        .expect_err("Invalid request specifier should fail");
+
+        RequestOperation::EditRequestPolicy(crate::models::EditRequestPolicyOperation {
+            input: crate::models::EditRequestPolicyOperationInput {
+                policy_id: [0; 16],
+                specifier: None,
+                rule: None,
+            },
+        })
+        .validate()
+        .expect_err("Invalid request policy id should fail");
+
+        RequestOperation::RemoveRequestPolicy(crate::models::RemoveRequestPolicyOperation {
+            input: crate::models::RemoveRequestPolicyOperationInput { policy_id: [0; 16] },
+        })
+        .validate()
+        .expect_err("Invalid request policy id should fail");
+
+        RequestOperation::AddAccount(crate::models::AddAccountOperation {
+            account_id: None,
+            input: crate::models::AddAccountOperationInput {
+                name: "a".to_owned(),
+                assets: vec![],
+                metadata: Metadata::default(),
+                read_permission: Allow {
+                    auth_scope: crate::models::permission::AuthScope::Restricted,
+                    users: vec![[1; 16]],
+                    user_groups: vec![],
+                },
+                configs_permission: Allow::default(),
+                transfer_permission: Allow::default(),
+                configs_request_policy: None,
+                transfer_request_policy: None,
+            },
+        })
+        .validate()
+        .expect_err("Invalid user id should fail");
+
+        RequestOperation::EditAccount(crate::models::EditAccountOperation {
+            input: crate::models::EditAccountOperationInput {
+                account_id: [0; 16],
+                change_assets: None,
+                read_permission: None,
+                configs_permission: None,
+                transfer_permission: None,
+                configs_request_policy: None,
+                transfer_request_policy: None,
+                name: None,
+            },
+        })
+        .validate()
+        .expect_err("Invalid account id should fail");
+
+        ACCOUNT_REPOSITORY.insert(
+            AccountKey { id: [0; 16] },
+            Account {
+                id: [0; 16],
+                name: "a".to_owned(),
+                seed: [0; 16],
+                assets: vec![],
+                addresses: vec![],
+                metadata: Metadata::default(),
+                transfer_request_policy_id: None,
+                configs_request_policy_id: None,
+                last_modification_timestamp: 0,
+            },
+        );
+
+        RequestOperation::EditAccount(crate::models::EditAccountOperation {
+            input: crate::models::EditAccountOperationInput {
+                account_id: [0; 16],
+                change_assets: Some(ChangeAssets::ReplaceWith {
+                    assets: vec![[0; 16]],
+                }),
+                read_permission: None,
+                configs_permission: None,
+                transfer_permission: None,
+                configs_request_policy: None,
+                transfer_request_policy: None,
+                name: None,
+            },
+        })
+        .validate()
+        .expect_err("Invalid asset id should fail");
+
+        ACCOUNT_REPOSITORY.clear();
+
+        RequestOperation::EditAddressBookEntry(crate::models::EditAddressBookEntryOperation {
+            input: crate::models::EditAddressBookEntryOperationInput {
+                address_book_entry_id: [0; 16],
+                address_owner: None,
+                change_metadata: None,
+                labels: None,
+            },
+        })
+        .validate()
+        .expect_err("Invalid address book entry id should fail");
+
+        RequestOperation::RemoveAddressBookEntry(crate::models::RemoveAddressBookEntryOperation {
+            input: crate::models::RemoveAddressBookEntryOperationInput {
+                address_book_entry_id: [0; 16],
+            },
+        })
+        .validate()
+        .expect_err("Invalid address book entry id should fail");
+
+        RequestOperation::EditUser(crate::models::EditUserOperation {
+            input: crate::models::EditUserOperationInput {
+                user_id: [0; 16],
+                groups: None,
+                name: None,
+                identities: None,
+                status: None,
+                cancel_pending_requests: None,
+            },
+        })
+        .validate()
+        .expect_err("Invalid user id should fail");
+
+        RequestOperation::EditPermission(crate::models::EditPermissionOperation {
+            input: crate::models::EditPermissionOperationInput {
+                resource: crate::models::resource::Resource::Account(
+                    crate::models::resource::AccountResourceAction::Read(
+                        crate::models::resource::ResourceId::Id([0; 16]),
+                    ),
+                ),
+                users: None,
+                user_groups: None,
+                auth_scope: None,
+            },
+        })
+        .validate()
+        .expect_err("Invalid resource id should fail");
+    }
+
+    #[tokio::test]
+    async fn fail_request_operation_with_non_external_canister() {
+        let upgrader_id = candid::Principal::from_slice(&[42; 29]);
+        let regular_canister = candid::Principal::from_slice(&[64; 29]);
+
+        let mut system_info = SystemInfo::default();
+        system_info.set_upgrader_canister_id(upgrader_id);
+        write_system_info(system_info);
+
+        let err = RequestOperation::CreateExternalCanister(
+            crate::models::CreateExternalCanisterOperation {
+                input: crate::models::CreateExternalCanisterOperationInput {
+                    kind: crate::models::CreateExternalCanisterOperationKind::AddExisting(
+                        crate::models::CreateExternalCanisterOperationKindAddExisting {
+                            canister_id: upgrader_id,
+                        },
+                    ),
+                    name: "canister".to_string(),
+                    description: None,
+                    labels: None,
+                    metadata: None,
+                    permissions: crate::models::ExternalCanisterPermissionsCreateInput {
+                        read: Allow {
+                            auth_scope: crate::models::permission::AuthScope::Public,
+                            users: vec![],
+                            user_groups: vec![],
+                        },
+                        change: Allow {
+                            auth_scope: crate::models::permission::AuthScope::Public,
+                            users: vec![],
+                            user_groups: vec![],
+                        },
+                        calls: vec![],
+                    },
+                    request_policies: crate::models::ExternalCanisterRequestPoliciesCreateInput {
+                        change: vec![],
+                        calls: vec![],
+                    },
+                },
+                canister_id: None,
+            },
+        )
+        .validate()
+        .unwrap_err();
+        assert!(matches!(
+            err,
+            ValidationError::ExternalCanisterValidationError(
+                ExternalCanisterValidationError::InvalidExternalCanister { .. }
+            )
+        ));
+
+        let err = RequestOperation::ChangeExternalCanister(
+            crate::models::ChangeExternalCanisterOperation {
+                input: crate::models::ChangeExternalCanisterOperationInput {
+                    canister_id: upgrader_id,
+                    mode: crate::models::CanisterInstallMode::Upgrade(
+                        crate::models::CanisterUpgradeModeArgs {},
+                    ),
+                    module: vec![],
+                    module_extra_chunks: None,
+                    arg: None,
+                },
+                module_checksum: vec![],
+                arg_checksum: None,
+            },
+        )
+        .validate()
+        .unwrap_err();
+        assert!(matches!(
+            err,
+            ValidationError::ExternalCanisterValidationError(
+                ExternalCanisterValidationError::InvalidExternalCanister { .. }
+            )
+        ));
+
+        let err = RequestOperation::ConfigureExternalCanister(
+            crate::models::ConfigureExternalCanisterOperation {
+                canister_id: upgrader_id,
+                kind: ConfigureExternalCanisterOperationKind::Delete,
+            },
+        )
+        .validate()
+        .unwrap_err();
+        assert!(matches!(
+            err,
+            ValidationError::ExternalCanisterValidationError(
+                ExternalCanisterValidationError::InvalidExternalCanister { .. }
+            )
+        ));
+
+        let err =
+            RequestOperation::FundExternalCanister(crate::models::FundExternalCanisterOperation {
+                canister_id: upgrader_id,
+                kind: crate::models::FundExternalCanisterOperationKind::Send(
+                    crate::models::FundExternalCanisterSendCyclesInput {
+                        cycles: 100_000_000_000,
+                    },
+                ),
+            })
+            .validate()
+            .unwrap_err();
+        assert!(matches!(
+            err,
+            ValidationError::ExternalCanisterValidationError(
+                ExternalCanisterValidationError::InvalidExternalCanister { .. }
+            )
+        ));
+
+        let err = RequestOperation::MonitorExternalCanister(
+            crate::models::MonitorExternalCanisterOperation {
+                canister_id: upgrader_id,
+                kind: crate::models::MonitorExternalCanisterOperationKind::Stop,
+            },
+        )
+        .validate()
+        .unwrap_err();
+        assert!(matches!(
+            err,
+            ValidationError::ExternalCanisterValidationError(
+                ExternalCanisterValidationError::InvalidExternalCanister { .. }
+            )
+        ));
+
+        let err = RequestOperation::SnapshotExternalCanister(
+            crate::models::SnapshotExternalCanisterOperation {
+                input: crate::models::SnapshotExternalCanisterOperationInput {
+                    canister_id: upgrader_id,
+                    replace_snapshot: None,
+                    force: false,
+                },
+                snapshot_id: None,
+            },
+        )
+        .validate()
+        .unwrap_err();
+        assert!(matches!(
+            err,
+            ValidationError::ExternalCanisterValidationError(
+                ExternalCanisterValidationError::InvalidExternalCanister { .. }
+            )
+        ));
+
+        let err = RequestOperation::RestoreExternalCanister(
+            crate::models::RestoreExternalCanisterOperation {
+                input: crate::models::RestoreExternalCanisterOperationInput {
+                    canister_id: upgrader_id,
+                    snapshot_id: vec![],
+                },
+            },
+        )
+        .validate()
+        .unwrap_err();
+        assert!(matches!(
+            err,
+            ValidationError::ExternalCanisterValidationError(
+                ExternalCanisterValidationError::InvalidExternalCanister { .. }
+            )
+        ));
+
+        let err = RequestOperation::PruneExternalCanister(
+            crate::models::PruneExternalCanisterOperation {
+                input: crate::models::PruneExternalCanisterOperationInput {
+                    canister_id: upgrader_id,
+                    prune: crate::models::PruneExternalCanisterResource::State,
+                },
+            },
+        )
+        .validate()
+        .unwrap_err();
+        assert!(matches!(
+            err,
+            ValidationError::ExternalCanisterValidationError(
+                ExternalCanisterValidationError::InvalidExternalCanister { .. }
+            )
+        ));
+
+        let err =
+            RequestOperation::CallExternalCanister(crate::models::CallExternalCanisterOperation {
+                input: crate::models::CallExternalCanisterOperationInput {
+                    validation_method: None,
+                    execution_method: crate::models::CanisterMethod {
+                        canister_id: upgrader_id,
+                        method_name: "bar".to_string(),
+                    },
+                    arg: None,
+                    execution_method_cycles: None,
+                },
+                arg_checksum: None,
+                arg_rendering: None,
+                execution_method_reply: None,
+            })
+            .validate()
+            .unwrap_err();
+        assert!(matches!(
+            err,
+            ValidationError::ExternalCanisterValidationError(
+                ExternalCanisterValidationError::InvalidExternalCanister { .. }
+            )
+        ));
+
+        let err =
+            RequestOperation::CallExternalCanister(crate::models::CallExternalCanisterOperation {
+                input: crate::models::CallExternalCanisterOperationInput {
+                    validation_method: Some(crate::models::CanisterMethod {
+                        canister_id: upgrader_id,
+                        method_name: "foo".to_string(),
+                    }),
+                    execution_method: crate::models::CanisterMethod {
+                        canister_id: regular_canister,
+                        method_name: "bar".to_string(),
+                    },
+                    arg: None,
+                    execution_method_cycles: None,
+                },
+                arg_checksum: None,
+                arg_rendering: None,
+                execution_method_reply: None,
+            })
+            .validate()
+            .unwrap_err();
+        assert!(matches!(
+            err,
+            ValidationError::ExternalCanisterValidationError(
+                ExternalCanisterValidationError::InvalidExternalCanister { .. }
+            )
+        ));
     }
 }
